+++ packages/expo-location/src/Location.ts
@@ -299,7 +299,7 @@ export const useForegroundPermissions = createPermissionHook({
  * Checks user's permissions for accessing location while the app is in the background.
  * @return A promise that fulfills with an object of type [`PermissionResponse`](#permissionresponse).
  */
-export async function getBackgroundPermissionsAsync(): Promise<PermissionResponse> {
+export async function getBackgroundPermissionsAsync(): Promise<LocationPermissionResponse> {
   return await ExpoLocation.getBackgroundPermissionsAsync();
 }
 
@@ -313,7 +313,7 @@ export async function getBackgroundPermissionsAsync(): Promise<PermissionRespons
  * (your app can't obtain background permission without foreground permission).
  * @return A promise that fulfills with an object of type [`PermissionResponse`](#permissionresponse).
  */
-export async function requestBackgroundPermissionsAsync(): Promise<PermissionResponse> {
+export async function requestBackgroundPermissionsAsync(): Promise<LocationPermissionResponse> {
   return await ExpoLocation.requestBackgroundPermissionsAsync();
 }